rename wear message
